pyproject.toml
[tool.poetry]
name = "dan-toropov-cw-5"
version = "0.1.0"
description = ""
authors = ["dan.toropov <dan.toropov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.9"
requests = "^2.32.3"
pytest = "^8.2.2"
coverage = "^7.5.4"
pytest-cov = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
